--- app/src/atomScope.jsx
@@ -1,33 +1,0 @@
-// src/app/src/atomScope.jsx
-// This file is maintained for backward compatibility
-import React, { createContext, useContext } from "react";
-import { usePythonAtomScope } from "./components/python/atomScope.jsx";
-import { useJSAtomScope } from "./components/js/atomScope.jsx";
-
-// Create a global context that will be used to determine which implementation to use
-const ImplementationContext = createContext("python"); // Default to python
-
-// Export the hook to access the current atom scope
-export function useAtomScope() {
-  const implementation = useContext(ImplementationContext);
-  
-  // Determine which implementation to use
-  if (implementation === "js") {
-    return useJSAtomScope();
-  } else {
-    // Default to Python
-    return usePythonAtomScope();
-  }
-}
-
-// Provider component to specify which implementation to use
-export const ImplementationProvider = ({ implementation = "python", children }) => {
-  return (
-    <ImplementationContext.Provider value={implementation}>
-      {children}
-    </ImplementationContext.Provider>
-  );
-};
-
-// Re-export the provider for backward compatibility
-export { PythonMolViewSpecApp as MolViewSpecApp } from "./components/python/atomScope.jsx";
